Removed unnecessary quotes around {{action}} name.
Considering the quotes around the action name are no longer necessary, I
think it'd be less distracting if we just presented them without the
quotes. Seems more intuitive this way: an action name ultimately
resolved to a function, which makes it kind of weird to represent as a
string when you're first learning this stuff.