Show timestamp in Blog post pages
When reading the blog, it is confusing not to see the date when a
post was written. For instance the following post:

http://emberjs.com/blog/2013/05/03/ember-data-progress-update.html

starts with this text "Just over a month ago, we told you about".
Having the timestamp at the top of the post can be helpful to
understand exactly what the post is talking about