Use border, not outline, for left nav
Using `outline` causes some of the `li`s to appear with a double border.
This commit replaces the `outline` with a `border-top` and
`border-bottom` and uses `margin-top: -1px` to collapse the two on all
but the first `li`.

I had initially tried adding the `border-top` only for the first `li`,
but that meant that when the inner levels were expanded, there was no
border between the nth inner content and the n+1st main nav.