Modify top section to use buttons instead of info.
1) Added two buttons (for a total of three) to initiate downloading
Ember.js.
2) Removed the info line that was below the single button before.
3) Change the word "debug" to "development".

I always thought that it was "kinda" confusing that the one button
didn't really say which version you would be downloading if you clicked
the button.

Change #3 may be backed out by the maintainers of Ember if they believe
the title "debug" is the accepted title for this but again, I think it
is "misleading" in that it sounds like "Download this version if you are
having problems and need to debug" when it really is "Use this version
for application development because it has additional functionallity to
help you debug your code if you need to".