Route.redirect is deprecated, replacing by afterModel
`redirect:` is not working with nested routes right now. I was digging the documentation and noted it was deprecated. I could not find why, so I did a blame and found this out:

https://github.com/emberjs/ember.js/commit/5e7a2636bf9e842b8a9c64d291c6408f3fab0eab

Using `afterModel` instead of `redirect` makes it possible to work with nested routes. I don't know why the auto generated API docs don't mention `afterModel` anymore, but this seems to fix the problem to me and it removes a deprecated method from the examples on the site.