Change redirection examples to use the `beforeModel` hook instead of the deprecated `redirect` hook.
The `redirect` route has been deprecated because it wasn't clear what state the model would be in at that point, making it harder to perform a redirect based upon the model's state. This was a change that landed with @machty's fully async transitions patch. More information on the API can be found here:
https://gist.github.com/machty/5723945