[WEBSITE GUIDES] Add New Features Page to Guides
It would be nice to have a comprehensive and direct list of features added as well as well as the release that they were added.

Not all users have the time to pore over the changelog and also hunt down descriptions of the feature and when it became available.

This gives people a single place to look for features on the website that is more visible.

TODO: add ember releases when these specific features became available.

Also there seems to be some drift between FEATURES.md and the CHANGELOG.md in the ember js git repo. Those should be made consitent as well.

Perhaps the FEATURES.md document can be structured and written in such a way that it is an easy drop in replacement for the website content in this guide.