Wrap Async Code In Ember.run
Shouldn't these setters and method that result in a set be wrapped in Ember.run for correctness?  Might also need to look at the other pages, because it's not consistent on when you need to wrap in `Ember.run` e.g. they are wrapped in the examples of "Testing Components" but not in "Testing Controllers".  Willing to do the work just need to verify these are the correct changes.