Fix sidebar for "Understanding Ember.js"
Since `lib/toc.rb` checks whether the first element of a chapter matches the current URL, having `deprecations` as the first item in the `understanding-ember` chapter caused problems with the sidebar. Namely, when accessing e.g. http://emberjs.com/guides/understanding-ember/the-view-layer/ nothing in the sidebar is selected.

Fixed by moving Deprecations to it's own chapter. This causes the least friction (as opposed to promoting Deprecations to "Understanding Ember.js"), since a lot of places link directly to http://emberjs.com/guides/deprecations