Update testing-routes.md
This file was no longer loading in the Live Example because it was removed in the `master` branch. 
  <script src="https://rawgithub.com/rpflorence/ember-qunit/master/dist/globals/main.js" ...

I have updated the link to point to the last release tag (current) which contains the file:
<script src="https://rawgithub.com/rpflorence/ember-qunit/v0.1.8/...

Upon saving a new JSbin ID was created. I've locked it for changes and am submitting as the new Live Example.