Remove the San Juan, PR meet up
This meetup slowed down and finally became
inactive. I wrote a blog post about the reasons
but in summary: our island is too small for a 
framework/language specific meetup.

We started a new meetup aimed at the general
developer community an it seems to be doing a lot
better. There's a lot of interest about Ember.js
so we'll keep doing talks and hand on learning
events.